Отредактировал миграции create_games_table, create_platforms_table, create_game_genre_table, create_baskets_table, create_favorites_table. Создал миграцию link_platforms_with_games

+++ database/migrations/2023_05_21_103453_create_game_genre_table.php
@@ -16,8 +16,8 @@ class CreateGameGenreTable extends Migration
         Schema::create('game_genre', function (Blueprint $table) {
             $table->id();
             $table->timestamps();
-            $table->unsignedInteger('game_id');
-            $table->unsignedInteger('genre_id');
+            $table->unsignedBigInteger('game_id');
+            $table->unsignedBigInteger('genre_id');
 
             $table->unique(['game_id', 'genre_id']);
             $table->foreign('game_id')->references('id')->on('games');
